split visual techniques in different py modules
