+++ clayrs/content_analyzer/field_content_production_techniques/visual_techniques/__init__.py
@@ -1,2 +1,3 @@
-from .visual_content_techniques import SkImageHogDescriptor, SkImageCannyEdgeDetector, SkImageSIFT, ColorsHist, \
-    SkImageLBP, ColorQuantization, PytorchImageModels
+from .low_level_techniques import SkImageHogDescriptor, SkImageCannyEdgeDetector, SkImageSIFT, ColorsHist, \
+    SkImageLBP, ColorQuantization, CustomFilterConvolution
+from .high_level_techniques import PytorchImageModels
